Makefile: Add removed target
Lets you list out what files were removed from OLD to NEW generations. Note
automatically done to remove files as I consider that the job of the user.

While here, added a function check_defined to let me check if a variable is
defined or not and if not bail.

Updated the option target to use this function as well as the removed target
where I needed it for.